Add files via upload
Simple program to check if a given numerical input is a valid Mastercard, American Express, or VISA credit card. I used Luhn's algorithm to mathematically check to see if the number could be a valid generated card number. I purposely did this without using arrays/lists/etc. This does not check for any malicious card numbers that pass the checksum in the mod 10 algorithm.